refactor: address PR review feedback
- Remove NextStepsSection wrapper (maxgfr feedback) — inline NextStepsBox directly
- Move UpdateCseModal from steps/step6/ to shared/ (fix inverted dependency)
- Create dedicated NextStepsBox.module.scss (fix shared SCSS import from steps/)
- Clean up unused styles from Step6Review.module.scss
